type casae

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -20,7 +20,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 		if (separator != NULL)
 		{
 			printf("%d", va_arg(list, int));
-			if (i != n - 1)
+			if (i != (int)n - 1)
 				printf("%s", separator);
 		}
 	}
